fix(0.9.0): API stabilization — event_id() returns typed EventId not u128 (#62)
Change the event_id accessors/fields that leaked raw u128 to the typed
EventId: IndexEntry::event_id, CommitMetadata::event_id + Committed::event_id,
the Notification.event_id field, and the CanalItem::event_id trait method.
Add a const EventId::from_u128 to the id macro so the const accessors can wrap.

StoreError::NotFound(u128) -> NotFound(EventId); AncestryCorrupt{cycle_at:u128}
-> {cycle_at: EventId}; Display arms format .as_u128() so hex output is
unchanged.

Fix every call site the compiler flagged, converting at honest boundaries with
EventId::from/.as_u128() and using .as_u128() at sort/compare sites (EventId is
not Ord). No allow/unwrap. Drop now-useless EventId::from(_.event_id())
conversions across crates, tests, benches, and examples.

+++ bpk-lib/crates/core/examples/wait_for_durable.rs
@@ -29,7 +29,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let entry = store
         .query(&Region::entity("ledger:acct-1"))
         .into_iter()
-        .find(|entry| entry.event_id() == u128::from(receipt.event_id))
+        .find(|entry| entry.event_id() == receipt.event_id)
         .ok_or("appended event missing from query")?;
     let target = HlcPoint {
         wall_ms: entry.wall_ms(),
